Module ruststep::ap000

Manually generated schema definitions corresponding following EXPRESS Schema

SCHEMA ap000;
  ENTITY a;
    x: REAL;
    y: REAL;
  END_ENTITY;

  ENTITY b;
    z: REAL;
    w: a;
  END_ENTITY;

  ENTITY c;
    p: a;
    q: b;
  END_ENTITY;

  -- For subtype/supertype
  ENTITY base;
    SUPERTYPE OF (sub)
    a: f64;
  END_ENTITY;

  ENTITY sub;
    SUBTYPE OF (base);
    b: f64;
  END_ENTITY;

  ENTITY user;
    data: base;
  END_ENTITY;
END_SCHEMA;

This sub-module is for help designing and testing generated code. Most functionality in generated codes are supplied as trait in tables.

Examples

use ruststep::*;

const STEP_INPUT: &str = r#"
ISO-10303-21;
HEADER;
  FILE_DESCRIPTION((''), '');
  FILE_NAME('ruststep/examples/ap000/read.step', '2018-04-27T08:23:47', (''), (''), '', '', '');
  FILE_SCHEMA(('AP000'));
ENDSEC;
DATA;
  #1 = A(1.0, 2.0);
  #2 = B(3.0, #1);
  #3 = B(3.0, A((4.0, 5.0)));
  #4 = C(#1, #2);
  #5 = C(#1, #3);
  #6 = C(#1, B((6.0, #1)));
  #7 = C(#1, B((6.0, A((7.0, 8.0)))));
  #8 = C(A((9.0, 10.0)), #2);
  #9 = C(A((11.0, 12.0)), #3);
ENDSEC;
END-ISO-10303-21;
"#;

// Parse input string into an exchange structure
let step = parser::parse(STEP_INPUT.trim()).unwrap();

// STEP file can contain multiple DATA section,
// and assumes it be 1 here.
assert_eq!(step.data.len(), 1);

// Load DATA section as tables of each entity
let table = ap000::Ap000::from_section(&step.data[0]).unwrap();

// Iterate over entity instances
for c in table.c_iter() {
    let c_owned = c.unwrap(); // Entity reference e.g. `#1` is resolved here.
                              // If an undefined entity is contained, `c` will be
                              // `ruststep::error::Error::UnknownEntity`
    println!("C = {:?}", c_owned);
}

custom Any trait for entity a

use ruststep::ap000::*;

let base = Base { a: 1.0 };
let sub = Sub { base, b: 1.0 };

let sub_r = &sub as &dyn BaseAny;

// call Debug for Sub by dispatch
dbg!(&sub_r);

let sub2: &Sub = sub_r.downcast_ref().unwrap();

Structs

A
Corresponds to ENTITY a

AHolder
Holder for A

Ap000
Tables including entities A, B, and C as their holders.

B
Corresponds to ENTITY b

BHolder
Holder for B

Base

C
Corresponds to ENTITY c

CHolder
Holder for C

Sub

User

Traits

BaseAny