Module ruststep::header

Header section of exchange structure

HEADER section of exchange structure shall contains one instance of each of following entities in this order:

file_description

file_name

file_schema

and following entities may appear after file_schema:

schema_population

file_population

section_language

section_context

These entities are defined in ISO-10303-21 “8.2 Header section declarations” using EXPRESS schemas. Although we can generate corresponding Rust struct using espr compiler, we write these definitions manually to keep development process simple.

Structs

FileDescription

File description

FileName

File name

FileSchema

File schema

Header

STEP-file HEADER section