Trait ruststep::tables::Holder

pub trait Holder: Clone + 'static {
    type Owned;
    type Table;
    fn into_owned(self, table: &Self::Table) -> Result<Self::Owned, Error>;
}

Trait for resolving a reference through entity id

STEP input is once deserialized into a holder struct, e.g. ap000::AHolder, which contains PlaceHolder for skipping table lookup. Entity table represented by the trait EntityTable will be looked up to resolve the reference at PlaceHolder::into_owned, and this trait is for supporting this method.

Holder structs, e.g. ap000::AHolder will be generated by espr, and implementation of this trait will be also generated at the same time.

Associated Types

type Owned

type Table

Loading content...
Required methods

fn into_owned(self, table: &Self::Table) -> Result<Self::Owned, Error>

Loading content...