Crate ruststep

Reference of ruststep, a STEP toolkit written in pure Rust

See README.md and ARCHITECTURE.md on GitHub for the overview and high-level architecture of this project. This reference focus on the detail usage of this crate.

Re-exports

pub use derive_more;

Modules

ap201

ap203

ast

AST (abstract syntax tree) for exchange structure (ISO-10303-21)

error

header

parser

Tokenize ASCII exchange structure (or STEP-file) defined by ISO-10303-21

place_holder

primitive

Primitive types appears in STEP and not defined in Rust

tables

Traits for espr-generated structures

Macros

as_holder

Get Holder struct identifier from ENTITY struct identifier

Derive Macros

Deserialize

Generate impl Deserialize for entity structs

Holder

Generates Holder struct and related implementation for each ENTITY struct

TableInit

Derive TableInit for tables