Module ruststep::ast

AST (abstract syntax tree) for exchange structure (ISO-10303-21)

Serde data model

Parameter and Record can be deserialize through serde data model.

Parameter	serde data model

Integer	i64

Real	f64

String	string

Enumeration	-

List	seq

NotProvided	unit

Omitted	unit

Typed, Record	map

RValue	enum

See the official document of serde data model for detail.

Parameter::Typed e.g. A((1.0, 2.0)) and Record e.g. A(1.0, 2.0) are mapped to “map” in the serde data model

Parameter::RValue is mapped to “enum” in the serde data model.

FIXME Enumeration is not supported yet.

Structs

Anchor

DataSection

DATA section in STEP file

Exchange

Entire exchange structure

Record

A struct typed in EXPRESS schema, e.g. A(1.0, 2.0)

ReferenceEntry

SingleMapDeserializer

Deserializer corresponding to a single-key map like { "A": [1.0, 2.0] }

URI

Enums

AnchorItem

EntityInstance

Each line of data section

LValue

Left hand side value

Parameter

Primitive value type in STEP data

RValue

Right hand side value

Functions

to_record

Serialize struct into STEP Record