Module ruststep::ast

AST (abstract syntax tree) for exchange structure (ISO-10303-21)

Deserialize Parameter into Rust struct

Parameter and Record can be deserialize through serde data model.

Parameter	serde data model

Integer	i64

Real	f64

String	string

Enumeration	-

List	seq

NotProvided	unit

Omitted	unit

Typed	map

RValue	enum

See the official document of serde data model for detail.

Parameter::Typed is mapped to “map” e.g. A((1.0, 2.0)) will be deserialized into { "A": [1.0, 2.0] }.

Parameter::RValue is mapped to “map” e.g. an entity reference #12 will be deserialized into { "Entity": 12 }.

FIXME: Enumeration is not supported yet.

Examples

Untyped parameters, e.g. real number

use serde::Deserialize;
use ruststep::ast::Parameter;

#[derive(Debug, Deserialize)]
struct A {
    x: f64,
    y: f64,
}

// Create a list as `Parameter::List`
let p = Parameter::from_iter(&[Parameter::real(1.0), Parameter::real(2.0)]);

// Deserialize the `Parameter` sequence into `A` because serde allows upcasting seq to struct
let a: A = Deserialize::deserialize(&p).unwrap();
println!("{:?}", a);

// Input types will be checked at runtime:
let p = Parameter::from_iter(&[Parameter::string("a"), Parameter::integer(2)]);
let result: Result<A, _> = Deserialize::deserialize(&p);
assert!(result.is_err());

Typed parameter, e.g. A(1)

FIXME: Type name check is not implemented yet.

use serde::Deserialize;
use ruststep::parser::exchange;
use nom::Finish;

#[derive(Debug, Deserialize)]
struct A {
    x: f64,
    y: f64,
}

let (res, p) = exchange::parameter("A((1.0, 2.0))").finish().unwrap();
assert_eq!(res, "");
let a: A = Deserialize::deserialize(&p).unwrap();
dbg!(a);

For RValue

use serde::Deserialize;
use ruststep::{parser::exchange, ast::RValue};
use nom::Finish;

let (res, p) = exchange::parameter("#11").finish().unwrap();
let a: RValue = Deserialize::deserialize(&p).unwrap();
assert_eq!(a, RValue::Entity(11))

Structs

Anchor

DataSection
DATA section in STEP file

Exchange
Entire exchange structure

Record
A struct typed in EXPRESS schema

ReferenceEntry

SingleMapDeserializer
Deserializer corresponding to a single-key map like { "A": [1.0, 2.0] }

URI

Enums

AnchorItem

EntityInstance
Each line of data section

LValue
Left hand side value

Parameter
Primitive value type in STEP data

RValue
Right hand side value

Functions

to_record
Serialize struct into STEP Record