Module ruststep::parser::combinator

Parser combinators extended for STEP exchange structure

This is helper submodule for writting a parser like as WSN definitions.

Token separators in exchange structure is one of

space

explicit print control directives (\N\ and \F\ )

comments

and combinators in this submodule responsible for handling them.

Traits

ExchangeParser
Alias of nom::Parser

Tuple
helper for tuple_

Functions

char_

comma_separated

comment
Comment

ignorable

many0_

many1_

opt_

separated

separator
Comments with front/back spaces, or multi-space at least 1 char

tag_

tuple_
Sequence of separated tokens

Type Definitions

ParseResult
Parse result